VB working now for both single and multivariate Gaussian.
For scalar Gaussians the error was simply in the test: the learner
was initalized to the correct priors for the emissions, so its elbo correctly
didn't make any progress. This was misdiagnosed as a problem.

For multivariate Gaussians, the error was that the expected sufficient
statistics were being handed to the natural parameter functions, which was
incorrect.

Still need to fix some design issues with svihmm/models/State.py